backend/notifications/requirements.txt
psycopg2-binary>=2.9.0
PyJWT>=2.8.0
pywebpush>=1.14.0
